feat(monitor): 청산 조건 ④ 선물 급락 데이터 수신 구현 (Phase α-0 항목 2)
명세: '종목의 최고점 시각에서 현재까지 선물이 1% 하락 시 전량 청산'

[변경]
- kis_api/kis.py: KOSPI200 선물 (101S3000) WebSocket 구독 메서드 subscribe_futures() 신규
  H0IFCNT0 TR ID 파싱 추가 (float), 재접속 시 선물 자동 재구독
- main.py: run() 초반에 선물 구독 호출, _on_futures_price 콜백 신규
  → monitor.on_futures_price() 연결 (이전 754-755 TODO 자리)
- monitor.py: _futures_price, _futures_at_high 타입 int → float
  on_futures_price(price: float), 로그 포맷 :.2f
  (KOSPI200 선물가는 소수점 포함 — 350.25 같은 값)
- stock.py: futures_price_at_high 타입 int → float
- tests/test_futures_exit.py: 단위 테스트 10건 신규

[이전 → 이후]
이전: 청산 조건 ④ 로직만 있고 선물 데이터 미수신 → 영구 무력 상태
      선물가 타입 int → 소수점 손실
이후: KIS WebSocket으로 선물 데이터 수신 + float 타입 정확한 비교
      5개 청산 조건 모두 작동

Phase α-0 항목 2 완료. ISSUE-025 해결.

+++ src/models/stock.py
@@ -55,7 +55,7 @@ class TradeTarget:
     high_confirmed: bool = False                  # 고가 확정 (1% 하락 트리거)
 
     # ── 선물 추적 ──
-    futures_price_at_high: int = 0               # 종목 고점 시각의 선물 가격
+    futures_price_at_high: float = 0.0            # 종목 고점 시각의 선물 가격 (소수점 포함)
 
     # ── 매수 주문 ──
     buy1_placed: bool = False                    # 1차 매수 주문 접수 여부
